Exit non-monadic interface
